+++ build_runner/lib/src/watcher/graph_watcher.dart
@@ -59,7 +59,7 @@ class PackageGraphWatcher {
                   'Error from directory watcher for package:${w.node.name}\n\n'
                   'If you see this consistently then it is recommended that '
                   'you enable the polling file watcher with '
-                  '--use-polling-watcher.',
+                  '--use-polling-watcher.\n\n',
                   e,
                   s);
             }))